requirements.txt
openai
python-dotenv
mistralai
python-decouple
matplotlib
numpy
anthropic
streamlit
pillow
joblib
sqlalchemy
alembic
psycopg2-binary